feat(status-bar): change type of tip in State
